"2:40pm. I am back. I actually forgot to have breakfast while I was in bed so I am having it now. Let me finish it and then I will note my thoughts.
2:45pm. Let me read Youjo Senki and Sengoku Komachi. This is what my morning would usually be like, but it is the middle of the day.

Anyway it is all or nothing here. If I log into the channel now and see my post has been removed, not much I can do about that. I'll spend the next few years writing. If they are asking questions about the language, I am gold. I'll be able to proceed on my path as a programmer.

I can't really commit to writing until I resolve this issue pressing down on me. Let me do my reading and then I will take a look.

3:15pm. Whatever, let me check the channel. At this rate I'll just get nervous. I wasn't nervous before, but once you get the urge and suppress it, it turns to nervousness. There is no need to make a big deal out of this. Let's see what cards will I get dealt.

Nothing yet. The post is still there, but no replies have come to it. It seems the place is really slower than I thought. Nobody looked at my post yet.

Anyway...this is pretty anti-climactic. I'll let that bait hang there. I've time to think about what my moves should be at Tenstorrent and my future proposals for it, so there is no need to dwell upon it too much.

Actually, some ideas have come to me even now. I thinking what I could say to the CEO as potential business opportunities. Pretty much everything that involves optimization is a target. I'd suggest to him that once the hardware + software foundation is there, that he hire software people to explore those possibilities and make products that can be sold. There is no need to rely purely on hardware.

They are aiming to revolutionize computing, so there is no need to focus too much on the narrow game of deep learning.

3:30pm. Remember how those two posts hung on Reddit for months with no replies until I got rid of them recently?

I think their Discord channel is a much better place seek attention that that out of the way Reddit channel that was started by somebody unrelated to the company, so it is unlikely that I will go without a reply for too long. But for whatever reason, I can't wait forever here. Am I going to lie in bed all day waiting for the fish to come? No. I am going to have to start writing.

3:35pm. Yeah, this is perfect. In fact, they did open the channel last year in September, I got the months wrong by about 4. So yes, because I was not paying attention I lost a big opportunity to showcase Spiral much earlier. But I can correct it like this. It is important to give it a try. Maybe they'll go for it, and if that is the case it would have been a tragedy to miss the chance.

3:50pm. Let me chill a bit, maybe take a bath and I will close it early for the day. I am going to start focusing on the writing grove. Write the pages and cultivate readership. I need to put this piece of bait out of mind. If I do not get a reply by tomorrow, it is going to be questionable when it is going to arive if at all. I am only assuming that those people are paying attention to their Discord, but it is pretty bereft of activity. Most of the messages are hellos from when the people first joined.

They are US based so a day or two should be enough for them to see it and think about it. If I do not get an answer by tomorrow, the chances will be a lot higher that it will get neglected. If two days pass, I can assume the interest is cool. In that case I'll go to the Discord channel and just leave my email there instead of checking it daily. Or maybe I will just tell them to open an issue on the Spiral repo. Instead of actually leaving the email I'll just note that it is on my Github profile.

4pm. Now...it is chilling time. I need to finish sorting out my emotions and focus determinately on writing. I've gone through the Tenstorrent scenarios in my head, so I need to leave that behind. I can't depend on anything. I just have to do what I want.

4:05pm. Let me close here. Tomorrow I am going to get some writing done. Until I have access to the Tenstorrent chips I'll pretend it does not exist."