"4:50pm. Done with lunch. Let me go to bed again.
8:05pm. I was in bed this whole time thinking. Yeah, it is not easy at all to pivot in life. Tomorrow, I'll log into the channel and see where I stand, but my luck is horrible so who knows. I do not have forever to wait for them. I am not really a mercenary guy that juggles his options. I like being fixated on one goal only.

I am starting to get into it. Into the possibilities of writing I mean. If I were to write a regular story it is unlikely to be popular, but if I give it religious connotations throughout, it will be good. I must not forget - most people haven't read the original Simulacrum and will be blow away by the concepts when I introduce them. I have no idea what I was thinking in 2014 when I just spammed Reddit for a bit to promote the story and then gave up. I think I posted it on Wattpad and Spacebattles, but that was halfhearted.

I have to go higher and higher, and present a viable alternative to 'serving humanity' the AI friendliness peddlers are passing around.

Last month was really weird. I really had fun doing the C backend, but the rest can take a hike. I finally found a good fishing spot, but I do not have the patience to wait for it.

Rather than writing the story like I have before, as a stream of consciousness from start to end, I am thinking of outlining it first. Make a bunch of chapter titles, and write the key bullet points for every step of the way. I didn't do this last time in 2014. But I should take some lessons from my programming experience and do it.

8:10pm. I need to give this a try. 2014 was a special thing, it was a project of pure passion. That I was going to publish the book was just an excuse to give me an extra push to write it.

8:15pm. I do not really feel like thinking pleasantries about my prospects with Tenstorrent, I'll forget the hope, login into the channel and leave the note to open an issue if they want to get in touch with me. This will be enough to fulfil my obligation to Spiral. I have put a lot of my life into it, but it is not my life and I can't be beholden to it forever.

It is time for the next thing. I should approach the writing project like a programming project. There is no need to just zoom forward without direction like in 2014. I should first set the landmarks and then make sure the tracks follow them.

But of course, the most important thing would be to establish my motivation to write it in the first place. Lying in bed all day, just agonizing about the future is a really good way of doing it.

It is not against my principles to do this. Sometimes you are unsure and the best thing to do is to go into isolation. Away from the screen, away from any distractions. That will refine the possibilities until only the dominant one remains. The problem only starts if you spend this time playing games instead, that will just get you addicted to them and never build up the strength to pursue a path.

There is one thing I am sure. If I want to succeed at this I need to be really shameless. But at the same time, I am sure I am really good at that.

I really am not good at talking as those recent interviews showed me, but I should be able to have some confidence in my writing. Would it be 3/5 at least? Or maybe 4/5? I never thought what the rank for it would be.

I think 5/5 is really deserved for those who can start their own cults. I do not fit the grade. 5/5 needs some kind of accomplishment. I have that in functional programming, I might as well aim for it in writing.

8:25pm. Hmmm, right. I need to mention art. I did put a bunch of effort into 3d, but since I am going to be using DALLE style nets in the future, I am going to have to look into how to doodle. Draw 2d on the level of OPM as drawn by ONE. I am sure that given a text description DALLE will be able to work miracles.

If not 2d, then I am sure 3d will work perfectly. So that is something to keep in mind. Even though I've put in 9 months into it, I am more of an ideas guy and do not feel like being a modeler. I could really fall in love with it if I could fork myself and solely dedicate that instance to the particular role, but just writing itself is enough for me right now. When it comes to art, I want quality as quickly as possible. I do not care if it is not my own handywork. It is my writing that will be irreplaceable, not anything else.

I simply do not have enough productive capacity for anything left after writing.

It is really a pity about music, but I do not even have the will to study that anymore. It really is a pity as I have some cool tracks in my head.

8:30pm. Tomorrow I am going to have to look up how long it usually takes for the DALLE waiting list to clear. I am not really worried about being rejected here, this is not a job application. In future there will be plenty of such services as well. Even though I've read that DALLE costs 0.1$ per image, these guys have zero moat and the cost is going to collapse before long.

That startup Neurolabs I applied to uses convolutional NNs to count shelf bottles and displays labels which also costs 0.1$ per image, but since it is a niche product they are less likely to have a competitor.

8:40pm. Tomorrow I am going to most likely lie in bed all day again, but maybe I can talk myself into doing the outlining.

I was wondering what my writing score would be, but you can improve it in practice by effective planning and sketching. If you aren't really sure about something just sketch it out. It works in programming when you do stumps of functions leaving the implementation for later. It will work in writing with hierarchical chapter structure and short descriptions of them.

Instead of just pure, frenzied passion like in 2014, why not architect the story ahead of time like with programming?

If I can complete that step, it will create a lot of motivation to do the rest."