"2:20pm. I was in bed for quite a while today. I'll go back as soon as I finish up here. I only got up to have breakfast.
i've accepted many things about me, but one thing I haven't covered yet is my sadness. I've been trying to so hard for so long and yet the world is only throwing hurles at me. I have to deal with its coldness with willpower and perseverance. But what do I do when I do not have the resources to deal with the challenges. When everything becomes left to luck and the dice do not land right?

What am I supposed to do?

Sometimes pushing past the sadness is not the right answer. Instead I can turn it inside out and ask the world for help. If it has to make me feel like a human, I should act like a human. Plenty of people are in the same hopeless situation. My cry will summon them, and they will become my power.

The grade of a writer is not determined by the intelligence of the writer. Plenty of people are smart and yet write forgettable stories.

A mark of a top class writer or an artist is not technical proficiency, but the power to grasp the heart of the readers. I will open up a path, and invite the readers onto the journey to omnipotence. Simply writing a story is not enough. Leaving it to luck is not enough. Heaven's Key succeeding won't be a matter of luck.

I won't let it. I have the power to do this in myself. If my lonely struggle is hopeless, then I will turn that hopelessness into hope.

I will look my defeat in the eye. And this time I will accept it an emphasize with the other people in such a situation.

2:35pm. Playing the same trick, willpower, does get old after a while. No way will the universe allow it. It will only allow me to come so far.

I know I said I was going to do the outline, but that does not matter and can be done later. What matters is that I put this profound sadness and desire for power into words. All this world is asking me is to put a bit of sincerity into my efforts.

Nobody cares about the sci-fi gizmos. The stuff I wrote last year was third rate. I am going to demonstrate what first rate writing is. The story needs to be an entry into a dream.

2:40pm. Let me go to bed here.

7:25pm. I am up. I feel like I've gathered a bit more motivation.

This time I do not want to do things as fiercely as before. I tried ML and it culminated in me wanting to ask the machine a question. Why don't I write with the intention of asking a question of the world? It won't be to make an announcement like last time.

It is a perfect fit for the me as I am now riddled with doubts and uncertainties.

Also forget the outline. Story writing is all about momentum.

I can't forget the most important thing, that I in fact got right last time.

For the story to work, it has to not just entrance the reader, but the writer itself.

The quality of a particular piece of code is how easy it is to reason about.

The quality of a piece of writing is the expressive energy contained in it. It is not good unless it steals the writer's soul.

Things like outline, story structure, characterization take a backseat to that.

The raw passion is the mark of a true first grade artist.

7:30pm. The most important part of the story is the very first one. It is what will make me motivated to write the rest. It is what will convince the reader to give it a try.

I am just as much as the target of the story I am writing as the reader is. I can't forget that. If I remove myself from the equation, the story will be second rate at best and I will be stuck doing a dull job.

7:35pm. Today, I've spent the day in sadness. I'll keep doing these naps until my system digests the emotions and my new motivation is established properly. There is no need to force myself to write. I should instead bring myself to the point where I do not feel like doing anything else other than write.

I am sad because if wanting to win just brought me to this point, just what was the point of it all? Was there a point to having such a strong desire to self improve? I want the world to give me an answer. Since I have the proclivity towards writing, going down that path is the best way to go.

Money is secondary. I'll ensure the proper system is set up for the readers to support me, but whether I end up making 10$ a month or 1k a month is all up to fate. If I can't win, I simply can't win. Deepmind or somebody else will make that friendly AI and we will all be suppressed to never access the self improvement loop. Humanity will be safe, and desire and ambition of the individual will perish. It will only be the fault of others for letting it come to that. I can honestly say I gave it my best.

Simply a job will not allow me to close the gap with those giant, well funded research groups. I need to ask the world a question - just how does it want me to beat them? I need a sign.

I won't look for it from within anymore. The reaction to the story will be clue to home me in.

7:40pm. I should believe in myself. I said the passion is the mark of a first grade artist, but for most people that is probably the most difficult part to emulate. I can do it. Just how many times have I tricked myself into doing programming? I should be able to manage this for writing itself.

But I won't be as intense as before. I can only make proposals and observe the reaction to them.

7:45pm. It could be a fun journey. DALLE is something good. Maybe something further will come along those lines. Just by writing, I might be able to upgrade myself into a one man studio eventually. Maybe me getting rejected here will be a favor in disguise.

7:50pm. There is no need to be in a hurry to ask a question. There is no need to be intense in searching for an answer. I want to keep needling the world. Is it only going to keep sending me hurdle after hurdle or is it going to give me a little bit of what I need to win?"