"8:20am. Nothing of note in the mail.
At first I was fearful, then I was anxious, then I was sad, and towards the end of yesterday I became perplexed. That is how I am now.

8:35am. I am reflecting on my past mistake. Before I became a programmer I was a trader and I came at it with the same strong-willed, intense if lazy approach. The end result was failure. I learned to trade, but not so much how to make money.

I think maybe if I was less willful about it the results would have been different. With programming writing the early arc of Simulacrum inspired me to pursue the Singularity, but in the end what was the point of that if I am only going to come to the conclussion that I need to use the machine to derive the result?

If I got those AI chips now, sure, I'd have an edge due to having Spiral, but given enough time I am sure that the Deepmind guys will be able to do it themselves.

So what was the point? Do I have a single path of advancement left to me?

Freelancing could be different if there were different ranks to it. Like in those adventure guilds, you could start at the bottom taking F ranks, and then work your way up to something that pays well. But from what I've seen of Freelance and Upwork, all the work is F rank and I do not know if there is any chance of advancement, so that makes it very unappealing as a path.

Jobs are weird. Suppose programmers were wizards, and you have beginner programmers which can only cast lvl 1 spells all the way to the experts which can cast lvl 3,4,5.

I was better than my cohort in high school, but it is not like I have anyone to compare myself. I might be a lot better than my past me, but maybe I am just a frog in a well thinking he is lvl 5 when he is really lvl 3. So suppose I start applying and get rejected saying: 'Oh no, based on the level of expertise demonstrated you are just a level 3 wizard, we want somebody who is at least level 5.'

I could accept it. It would give me a goal and something to strive for.

But this scenario where I am being tested on my attitude and personality, as well as only tested on the bare minimum of my skill sits ill at ease for me. This kind of atmosphere turns me off and is making me unsure whether I even want a job to begin with. Forget lying on my resume and bullshiting for a moment.

Fundamentally, the job of the recruiters should be to gauge my skill accurately and hire me based on that. If it is not, I should find a place where it is.

Imagine if the application was for a neurosurgeon. You probably wouldn't care whether the guy was charming and personable, but whether he would kill the patient or not.

The job I want should test the extremity of my skill.

8:55am. So when I stopped programming last year, I went into art and cultivated my 3d skills, but it is not really my talent.

What would have made sense back then would be to start writing immediately.

My problem is that it is hard to have a strong attitude when you go into writing. Back in 2014, I figured out an important part of reality and was filled with pride, but right now I do not feel like I've discovered anything more noteworthy. Sure based on 7 years on programming I could put some twists to get closer to the truth, but it is not fundametally as earthshaking.

It is simply impossible for me to emulate the attitude of the 2014 early arcs. I have nothing to warrant such a level of will and confidence.

9:10am. But maybe that is the wrong reason to write.

The last thing I read was the Legendary Mechanic. It was fun for a bit, but it is a story about a robot of a human who only wins and ever grows stronger. It does not have much value as a literary work.

Right now, I am thinking and I am coming to the conclussion that this humbled, perplexed attitude might be the ideal way to approach the story. It will give it a sense of balance and contrast that would not be there otherwise. Plus writing with a strong attitude is very draining.

What if I wrote the same way I am keeping this journal?

9:15am. For the upkeep of the journal I wasn't ever straining my willpower. I was just airing my thoughts all along.

9:20am. I can't really approach writing saying my goal is to get 1k subscribers or something like that. I have zero way of making that happen. It is not a game where I can improve by taking different choices and actions. It makes no sense to approach writing as if it were combat or sports. Writing is one place where it is obvious you are not competing based on technical proficiency. There aren't rules and they don't matter.

9:30am. Maybe I need to go back to the way I was during my school days. Coming to an understand of the self improvement loop burst forth a wave of desire in me to become a programmer, but back then I didn't know anything about it. I only had an idea that purpose of life is to attain omnipotence and was looking for a path to attain it that would be realistic. I must have spent over a decade nursing this deep seated belief and desire.

I do not hate that old me. I think that holding on to my belief back then was the only thing of worth to me.

9:35am. I am going to start writing, but first I want to digest this grief and perplexity. I'll reaffirm the strong desire in me, and while I am holding on to it, write on the side. I won't bother trying to put it into the work itself like back in 2014.

I cannot do without it.

A strong desire is like a light that illuminates your life.

9:50am. I need to change just a bit.

Let me do my morning reading instead of just standing here in thought and then I will go back to bed.

10:35am. Let me go to bed. I am going to leave catching up to OPM for later. I am going to try to internalize the new mindset and do some writing after that. Maybe I'll manage to find the strength tomorrow.

It is taking me a long time as writing is not a matter of determination. Doing Heaven's Key has to take place of writing this journal.

On the plus side, at least I haven't been sleeping poorly."