"9:35am. Got up 15m ago. I've read Kaiji, now let me read Tenken. I seems the anime for it is two months from now.
10:25am. Let me start. I have two goals right now.

* Figure out a better synopsis.
* Figure out how to get the formatting on RR under control without necessarily having to remove all the empty spaces my hand.
* Then actually publish the story.

10:30am. Let me focus on the synopsis.

> What would a person do if he had access to a computer with nearly limitless computational power?

Hmmm...

> In the post-Singularity era where limitless computation is cheaply available, there were legendary figures seeking to take advantage of that power and make it their own. Rising above humanity, they sought power without limit and succeeded.

10:45am. Hmmm, this is a bit better.

> A story of what games are really about.

> The journey of their self development are enshrined as stories.

> One such journey is through a simulated game of poker.

10:50am. Hmmmm...

> In the post-Singularity era where limitless computation is cheaply available, there were legendary figures seeking to take advantage of that power and make it their own. Rising above humanity, they sought power without limit and succeeded, leaving behind a story of their journey on the path of transcendence.

Better.

> With enough computational power, one can do anything. In an era where one can create whole universes with a push of button, one can have fun playing a game and optimizing one's own mind.

> Far in the sky, emanating golden radiance as it floats amongst the clouds, is the city made of gold! In its game rooms, poker games of life and death unfold! The clashes of chips and cards!

> Can you conquer the game and take that power into the real world?

11:05am. Whoever is playing me rolled well this time. These last 4 quotes have potential.

///

In the post-Singularity era where limitless computation is cheaply available, there were legendary figures seeking to take advantage of that power and make it their own. Rising above humanity, they sought power without limit and succeeded, leaving behind a story of their journey on the path of transcendence. With enough computational power, one can do anything. In an era where one can create whole universes with a push of a button, one can have fun playing a game and optimizing one's own mind. Can you conquer the game and take that power into the real world?

Far in the sky, emanating golden radiance as it floats amongst the clouds, is the city made of gold! In its game rooms, poker games of life and death unfold! The clashes of chips and cards!

Devouring all newcomers and existing for eons, the city remained grand and unshaken, until one day, a novice entered the game…

///

How about this?

Synopsis is all about attitide. I need to transmit expectations about the story without saying anything about it.

> they sought power without limit
> they sought power without limit and succeeded

Consider these two quotes in isolation. Just adding 'and succeeded' completely changes what the tone. It is not something people would expect in a story about rising above humanity, so it will be great bait.

NPCs can only write stories about humanity winning, so this is a great signal. I must signal my values in the synopsis, and the further they are from standard the better.

11:25am. I can't think of anything else to change. This is good enough. Let me move to publishing again. I need to figure out how to get the paragraphs under control.

https://www.royalroad.com/forums/thread/110270?page=1#pid947118

Nevermind this, how the hell do I translate between formating.

https://www.royalroad.com/forums/thread/102804

Why is this such a pain in the ass, what the fuck?

This is like the fifth thread I've looked into.

Ah, I see. If I paste it as plain text, then it gets rid of the superflous empty lines and gives me proper paragraphs. This is the solution and I found it by myself. Let me just roll with it.

> A side effect of the operation is that the brain will have slightly increased energy consumption which will require me to eat more if I use the link regularly.

Hmmmm, should I just get rid of this? Yeah, I think I should for consistency.

> It says the link isn’t powered by my biology, but the nanofog, so the use of it won’t increase my energy consumption. That is good.

I can only lose by making computation expensive so let me get rid of that.

I'll add Gore to content warning, but otherwise keep the tags as I have yesterday. Bad ends will have violent depictions, but otherwise it is not the focus of the story.

12pm. I submitted it. It says it can take up to 48h for one of their staff members to look at it. It is Tuesday, so I am safe on that front.

https://www.royalroad.com/profile/303912

Here is my profile: ghostlike. I wanted to take that on Wordpress back in 2015, but it was taken, so I added some numbers to make it go through.

12:05pm. Once it gets added to the site, I'll post the rest of the chapters as well as some of the illustrations that I've done. I read in a /g/ thread that Stable Diffusion will go out of beta next week.

12:10pm. Right now I am a bit in a daze.

Anyway, to get popular as a writer, I'll go with the slow burn method. I'll get the page count high and let people discover me on their own. Back in 2014, I had no choice but to spam Reddit to get some initial readers, but that is in the past.

Sure, with active marketing, you can get a lot more views, and faster, but I am not in a hurry. People recommending my work would be the best kind of advertizing I could ask for. The key is to find a niche.

Since there isn't really a story like Heaven's Key out there, that gives me a big edge.

12:20pm. Anyway, forget Royal Road for a few days. What I should focus on is writing the next chapter.

12:25pm. Let me have breakfast here."